0855 Save File overwrite check added for Pattern Discovery
git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench@5741 8b29bb6e-589a-435a-9a12-3ab907be3462

+++ components/discovery/src/org/geworkbench/components/discovery/view/PatternTableView.java
@@ -204,6 +204,16 @@ public class PatternTableView extends JPanel {
                     selected = new File(selected.getAbsolutePath() + ".regx");
                 }
 
+                if (selected.exists()) {
+					int o = JOptionPane.showConfirmDialog(null,
+
+					"Replace the file", "Replace the existing file?",
+							JOptionPane.YES_NO_CANCEL_OPTION);
+					if (o != JOptionPane.YES_OPTION) {
+						return;
+					}
+				}
+                
                 BufferedWriter writer = new BufferedWriter(new FileWriter(selected));
                 for (int i = 0; i < rows.length; i++) {
                     org.geworkbench.util.patterns.CSMatchedSeqPattern pattern = (org.geworkbench.util.patterns.CSMatchedSeqPattern) model.getPattern(i);
@@ -257,6 +267,17 @@ public class PatternTableView extends JPanel {
                         return;
                     }
                     File file = chooser.getSelectedFile();
+
+                    if (file.exists()) {
+    					int o = JOptionPane.showConfirmDialog(null,
+
+    					"Replace the file", "Replace the existing file?",
+    							JOptionPane.YES_NO_CANCEL_OPTION);
+    					if (o != JOptionPane.YES_OPTION) {
+    						return;
+    					}
+    				}
+                    
                     if (!file.getName().endsWith(".pat")) {
                         file = new File(file.getAbsolutePath() + ".pat");
                     }
